Format `CompressedTexture`, remove TODO
The new method of getting `CompressedTexture` extensions has the
additional benefit of getting all extensions that might be available.
Checking if the extension is supported beforehand would be messy;
`getSupportedExtensions` returns a `JSArray<JSString>` that would have
to have `dartify()` called on it (which isn't optimal). Meanwhile,
`getExtension` returns null if the extension isn't in the list of
supported extensions.