Implement asset manifests
Right now, the only way to remap URLs is via a single global asset map.
This is rather inflexible, and doesn't allow for optimizing asset loads.

Instead, implement a new "Asset Manifest" resource. This is a JSON file
that remaps URLs to different ones (for example, URLs with hashes). This
replaces the current `getUrlHash` method entirely. We can manually
create an AssetManifest, or load one via the ResourceManager and use that
to create a new ResourceManager with a different manifest.

While we're at it, change some behavior internally for handling
alternative webp assets so that we don't need global state or enforce
a hash format. There may be a better way of handling this in the future,
but that's not now.