Implement support for resource registries
Resource registries provide a standardized way to add support for
additional resource kinds with a different "kind" string. This allows
moving the management of these resources into the resource manager and
out of custom solutions.

The implementation attempts to maintain a single instance of a registry
for each registered kind on a ResourceManager. This allows adding these
to the ResourceManager via extensions. `add` is not considered part of
the ResourceRegistry due to the signature differing for each resource;
some have options and some don't take URLs but instead take values of
another sort.