Set provoking vertex to first
Modern APIs, such as Metal, Vulkan, Direct3D 12, use the first vertex
as the provoking vertex, whereas WebGL API uses last. This can cause
performance degradation as, behind the scenes, workarounds are done to
switch from first vertex to last vertex in the implementation.

Note this is only for flat shading, or solid colors.

It seems there is no reason not to set this for WebGL2.